pyproject.toml
[project]
name = "finlab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "groq>=1.0.0",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "qdrant-client[fastembed]>=1.17.0",
    "uvicorn>=0.41.0",
]
